chore(infrastructure): move mongodb client
