+++ src/kairos/shared/infrastructure/mongodb/client/client.ts
@@ -1,0 +1,6 @@
+import { mongodbConfig } from '@/config/mongodb';
+import { createClient } from '@/kairos/shared/infrastructure/mongodb/client/create-client';
+
+export const mongoDBClient = createClient({
+  uri: mongodbConfig.connectionString,
+});
