+++ src/kairos/shared/infrastructure/mongodb/client/create-client.test.ts
@@ -1,6 +1,6 @@
+import { createClient } from '@/kairos/shared/infrastructure/mongodb/client/create-client';
 import { type MongoClient } from 'mongodb';
 import { describe, expect, it, vi } from 'vitest';
-import { createMongodbClient } from '@/kairos/shared/persistence/mongodb/create-mongodb-client';
 
 describe('createMongodbClient', () => {
   it('should create a mongodb client with the provided connection string', () => {
@@ -8,7 +8,7 @@ describe('createMongodbClient', () => {
     const makeClient = vi.fn().mockReturnValue(expectedClient);
     const uri = 'mongodb://mongodb:27017/kairos';
 
-    const client = createMongodbClient({ uri, makeClient });
+    const client = createClient({ uri, makeClient });
 
     expect(makeClient).toHaveBeenCalledWith(uri);
     expect(client).toBe(expectedClient);
