+++ src/kairos/shared/infrastructure/mongodb/collection/users-collection.ts
@@ -1,4 +1,4 @@
-import { mongoDBClient } from '@/kairos/shared/persistence/mongodb/mongodb-composition';
+import { mongoDBClient } from '@/kairos/shared/infrastructure/mongodb/client/client';
 import { UsersCollection, usersCollectionName } from '@/kairos/shared/persistence/mongodb/users-collection-schema';
 
 export const usersCollection: UsersCollection = mongoDBClient.db().collection(usersCollectionName);
