--- src/kairos/shared/persistence/mongodb/mongodb-composition.ts
@@ -1,6 +1,0 @@
-import { mongodbConfig } from '@/config/mongodb';
-import { createMongodbClient } from '@/kairos/shared/persistence/mongodb/create-mongodb-client';
-
-export const mongoDBClient = createMongodbClient({
-  uri: mongodbConfig.connectionString,
-});
