+++ tests/__vitest__/integration-test.ts
@@ -1,6 +1,6 @@
 import { mongodbConfig } from '@/config/mongodb';
+import { mongoDBClient } from '@/kairos/shared/infrastructure/mongodb/client/client';
 import { ensureMongoIndexes } from '@/kairos/shared/persistence/mongodb/ensure-mongo-indexes';
-import { mongoDBClient } from '@/kairos/shared/persistence/mongodb/mongodb-composition';
 import { afterEach, beforeEach } from 'vitest';
 
 export function integrationTest(): void {
